initialize_SPI modification and confusion. (non)pre-compile changes.
There are many places where #if is used instead of if() in
the middle of functions (so, not above them) and I found a spot
where it stopped the program from doing what it was supposed to.
As a result, I think we should change all that make sense to
the if() form. This way our compiler finds errors better. (For
instance, there was a spot where initialize_SPI was given an
argument, even though the function is defined as void, and there
was no error.